backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.120.1"}
psycopg = {extras = ["binary"], version = "^3.2.12"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.11.0"
xrpl-py = "^4.3.0"
sqlalchemy = "^2.0.44"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
